src/api/pyproject.toml
[project]
name = "hackathon-ufmg-2026-api"
version = "0.1.0"
description = "FastAPI backend scaffold for the Hackathon UFMG 2026 project."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "duckdb>=1.2.2,<2.0.0",
    "faiss-cpu>=1.11.0,<2.0.0",
    "fastapi>=0.115.12,<1.0.0",
    "numpy>=2.2.5,<3.0.0",
    "openai>=1.76.0,<2.0.0",
    "pdfplumber>=0.11.6,<1.0.0",
    "python-multipart>=0.0.20,<1.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "pydantic-settings>=2.9.1,<3.0.0",
    "sqlalchemy>=2.0.40,<3.0.0",
    "uvicorn[standard]>=0.34.2,<1.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1,<1.0.0",
    "pytest>=8.3.5,<9.0.0",
]

[tool.uv]
package = false
